+++ packages/core/lib/config.ts
@@ -1,5 +1,14 @@
+import type { OnSpanEndCallbacks } from './batch-processor'
+import {
+  ATTRIBUTE_ARRAY_LENGTH_LIMIT_DEFAULT,
+  ATTRIBUTE_COUNT_LIMIT_DEFAULT,
+  ATTRIBUTE_STRING_VALUE_LIMIT_DEFAULT,
+  ATTRIBUTE_ARRAY_LENGTH_LIMIT_MAX,
+  ATTRIBUTE_COUNT_LIMIT_MAX,
+  ATTRIBUTE_STRING_VALUE_LIMIT_MAX
+} from './custom-attribute-limits'
 import type { Plugin } from './plugin'
-import { isLogger, isNumber, isObject, isPluginArray, isString, isStringArray, isStringWithLength } from './validation'
+import { isLogger, isNumber, isObject, isOnSpanEndCallbacks, isPluginArray, isString, isStringArray, isStringWithLength } from './validation'
 
 type SetTraceCorrelation = (traceId: string, spanId: string) => void
 
@@ -40,6 +49,10 @@ export interface Configuration {
   plugins?: Array<Plugin<Configuration>>
   bugsnag?: BugsnagErrorStatic
   samplingProbability?: number
+  onSpanEnd?: OnSpanEndCallbacks
+  attributeStringValueLimit?: number
+  attributeArrayLengthLimit?: number
+  attributeCountLimit?: number
 }
 
 export interface TestConfiguration {
@@ -120,6 +133,26 @@ export const schema: CoreSchema = {
     defaultValue: undefined,
     message: 'should be a number between 0 and 1',
     validate: (value: unknown): value is number | undefined => value === undefined || (isNumber(value) && value >= 0 && value <= 1)
+  },
+  onSpanEnd: {
+    defaultValue: undefined,
+    message: 'should be an array of functions',
+    validate: isOnSpanEndCallbacks
+  },
+  attributeStringValueLimit: {
+    defaultValue: ATTRIBUTE_STRING_VALUE_LIMIT_DEFAULT,
+    message: `should be a number between 1 and ${ATTRIBUTE_STRING_VALUE_LIMIT_MAX}`,
+    validate: (value: unknown): value is number => isNumber(value) && value > 0 && value <= ATTRIBUTE_STRING_VALUE_LIMIT_MAX
+  },
+  attributeArrayLengthLimit: {
+    defaultValue: ATTRIBUTE_ARRAY_LENGTH_LIMIT_DEFAULT,
+    message: `should be a number between 1 and ${ATTRIBUTE_ARRAY_LENGTH_LIMIT_MAX}`,
+    validate: (value: unknown): value is number => isNumber(value) && value > 0 && value <= ATTRIBUTE_ARRAY_LENGTH_LIMIT_MAX
+  },
+  attributeCountLimit: {
+    defaultValue: ATTRIBUTE_COUNT_LIMIT_DEFAULT,
+    message: `should be a number between 1 and ${ATTRIBUTE_COUNT_LIMIT_MAX}`,
+    validate: (value: unknown): value is number => isNumber(value) && value > 0 && value <= ATTRIBUTE_COUNT_LIMIT_MAX
   }
 }